Final Version

+++ src/main/java/ma/ac/uir/devmanagement/repository/ProjectManagerRepository.java
@@ -3,5 +3,9 @@ package ma.ac.uir.devmanagement.repository;
 import ma.ac.uir.devmanagement.entity.ProjectManager;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectManagerRepository extends JpaRepository<ProjectManager, Integer> {
+import java.util.List;
+import java.util.Optional;
+
+public interface ProjectManagerRepository extends JpaRepository<ProjectManager, Long> {
+
 }
